+++ packages/terra-table/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added `event` object to `onCellSelect` callback.
+
 ## 5.12.0 - (March 8, 2024)
 
 * Changed